[persistence,activity] db models moved from ya-persistance to ya-activity
