+++ core/activity/Cargo.toml
@@ -19,6 +19,7 @@ actix-rt = "1.0"
 anyhow = "1.0"
 chrono = "0.4"
 diesel = { version = "1.4", features = ["chrono", "sqlite", "r2d2"] }
+diesel_migrations = "1.4"
 env_logger = "0.7"
 futures = "0.3"
 lazy_static = "1.4"
