+++ core/activity/examples/activity_provider_api.rs
@@ -1,8 +1,7 @@
 use actix_web::{middleware, App, HttpServer};
 
-use ya_activity::service;
+use ya_activity::{db::migrations, service};
 use ya_persistence::executor::DbExecutor;
-use ya_persistence::migrations;
 use ya_service_api_interfaces::Provider;
 use ya_service_api_web::rest_api_addr;
