+++ core/activity/src/service.rs
@@ -1,7 +1,7 @@
-use crate::{api, provider};
-
 use ya_persistence::executor::DbExecutor;
 use ya_service_api_interfaces::{Provider, Service};
+
+use crate::{api, db::migrations, provider};
 
 pub struct Activity;
 
@@ -11,7 +11,9 @@ impl Service for Activity {
 
 impl Activity {
     pub async fn gsb<Context: Provider<Self, DbExecutor>>(ctx: &Context) -> anyhow::Result<()> {
-        provider::service::bind_gsb(&ctx.component());
+        let db: DbExecutor = ctx.component();
+        db.apply_migration(migrations::run_with_output)?;
+        provider::service::bind_gsb(&db);
         Ok(())
     }
