+++ core/persistence/Cargo.toml
@@ -13,7 +13,6 @@ anyhow = "1.0.26"
 bigdecimal = "0.1.0"
 chrono = { version = "0.4.6", features = ["serde"] }
 diesel = { version = "1.4", features = ["sqlite", "r2d2", "chrono"] }
-diesel_migrations = "1.4"
 dotenv = "0.14"
 libsqlite3-sys = { version = "0.9.1", features = ["bundled"] }
 log = "0.4.8"
